Move DB upsert|create to service
