+++ LinkUp/Services/Clients/ClientService.cs
@@ -1,4 +1,5 @@
 using ErrorOr;
+using LinkUp.Infrastructure.Data;
 using LinkUp.Models;
 using LinkUp.ServiceErrors;
 using LinkUp.Services.Clients.Interfaces;
@@ -8,9 +9,24 @@ namespace LinkUp.Services.Clients;
 public class ClientService : IClientService
 {
     private static readonly Dictionary<Guid, Client> _clients = new();
+
+    //NOT NEEDED IF IN CONTROLLER
+    private readonly AppDbContext _db;
+
+    public ClientService(AppDbContext db)
+    {
+        _db = db;
+    }
+    //ENDOF
+
     public ErrorOr<Created> CreateClient(Client client)
     {
         _clients.Add(client.Id, client);
+
+        //OR IN CONTROLLER
+        _db.Clients.Add(client);
+        _db.SaveChanges();
+        //ENDOF
 
         return Result.Created;
     }
@@ -36,6 +52,22 @@ public class ClientService : IClientService
     {
         var IsNewlyCreated = !_clients.ContainsKey(client.Id);
         _clients[client.Id] = client;
+
+        //OR IN CONTROLLER
+        var clientFromDBToUpsert = _db.Clients.Find(id);
+        if (clientFromDBToUpsert == null)
+        {
+            _db.Clients.Add(client);
+            _db.SaveChanges();
+        }
+        else
+        {
+            clientFromDBToUpsert.Name = client.Name;
+            clientFromDBToUpsert.Email = client.Email;
+            clientFromDBToUpsert.Password = client.Password;
+            _db.SaveChanges();
+        }
+        //ENDOF
 
         return new UpsertedClient(IsNewlyCreated);
     }
